Rename dynamic hook module to record
