+++ tests/test_record_hooks.py
@@ -1,4 +1,4 @@
-"""CPU contract tests for the additive dynamic hook path."""
+"""CPU contract tests for the additive record-hook path."""
 
 from __future__ import annotations
 
@@ -8,7 +8,7 @@ import pytest
 import torch
 
 from dmi.adapters.base import StepReservation
-from dmi.hooks.dynamic import (
+from dmi.hooks.record import (
     HookOutput,
     HookPointV1,
     HookSpecV1,
